fix(server): minor changes

+++ server/api/removeMeeting.js
@@ -4,9 +4,15 @@ const connection = require('../config/connection');
 
 router.post('/', async (req, res) => {
 	const { ID } = req.body;
-	const Q_Delete = 'DELETE FROM meetings WHERE `ID`=?';
-	const [result, ...rest] = await connection.query(Q_Delete, [ID]);
-	res.send(result);
+	if (ID === undefined) return res.status(400).json({ message: 'Invalid ID' });
+	try {
+		const Q_Delete = 'DELETE FROM meetings WHERE `ID`=?';
+		const [result, ...rest] = await connection.query(Q_Delete, [ID]);
+	}
+	catch(err){
+		res.status(500).json({ message: err });
+	}
+	res.status(200).json({ message: 'Removed meeting' });
 });
 
 module.exports = router;
